travel_ai/app/requirements.txt
# requirements.txt (修正版)
fastapi==0.115.12
graphviz==0.20.3
ipython==8.12.3
langchain==0.3.23  # 必须升级
langchain-core==0.3.51  # 必须升级
langchain-community==0.3.21  # 需要添加
langchain-openai==0.3.11
langgraph==0.3.20
langsmith==0.3.19
openai==1.68.2
pydantic==2.10.6
pydantic-settings==2.8.1
PyJWT==2.10.1
python-dotenv==1.1.0
redis==5.2.1
uvicorn==0.34.0
aiomysql==0.2.0
loguru==0.7.2
dnspython==2.7.0
iniconfig==2.1.0
pluggy==1.5.0
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
aiosqlite==0.21.0
pybreaker==1.3.0
ratelimit==2.2.1
langchain-tavily==0.1.5
mypy==1.15.0
langchain_chroma

